src/requirements.txt
python-memcached
boto==2.2.2
